+++ src/ChebyMap.cc
@@ -103,12 +103,12 @@ AstChebyMap *ChebyMap::_makeRawChebyMap(ndarray::Array<double, 2, 2> const &coef
         throw std::invalid_argument(os.str());
     }
     if (has_fwd) {
-        detail::assertEqual(lbnd_f.size(), "lbnd_f size", nin, "number of input axes");
-        detail::assertEqual(ubnd_f.size(), "ubnd_f size", nin, "number of input axes");
+        detail::assertEqual(lbnd_f.size(), "lbnd_f size", static_cast<std::size_t>(nin), "number of input axes");
+        detail::assertEqual(ubnd_f.size(), "ubnd_f size", static_cast<std::size_t>(nin), "number of input axes");
     }
     if (has_inv) {
-        detail::assertEqual(lbnd_i.size(), "lbnd_i size", nout, "number of output axes");
-        detail::assertEqual(ubnd_i.size(), "ubnd_i size", nout, "number of output axes");
+        detail::assertEqual(lbnd_i.size(), "lbnd_i size", static_cast<std::size_t>(nout), "number of output axes");
+        detail::assertEqual(ubnd_i.size(), "ubnd_i size", static_cast<std::size_t>(nout), "number of output axes");
     }
 
     return reinterpret_cast<AstChebyMap *>(astChebyMap(nin, nout, ncoeff_f, coeff_f.getData(), ncoeff_i,
@@ -139,8 +139,8 @@ AstChebyMap *ChebyMap::_makeRawChebyMap(ndarray::Array<double, 2, 2> const &coef
            << ", which is too short; length = nout + 2 and nout must be > 0";
         throw std::invalid_argument(os.str());
     }
-    detail::assertEqual(lbnd_f.size(), "lbnd_f size", nin, "number of input axes");
-    detail::assertEqual(ubnd_f.size(), "ubnd_f size", nin, "number of input axes");
+    detail::assertEqual(lbnd_f.size(), "lbnd_f size", static_cast<std::size_t>(nin), "number of input axes");
+    detail::assertEqual(ubnd_f.size(), "ubnd_f size", static_cast<std::size_t>(nin), "number of input axes");
 
     return reinterpret_cast<AstChebyMap *>(astChebyMap(nin, nout, ncoeff_f, coeff_f.getData(), 0, nullptr,
                                                        lbnd_f.data(), ubnd_f.data(), nullptr, nullptr,
